Corrected some import to partial import.
Also corrected some mpi bug in evaluator.py
Note that individuals are created with no reference to mpi.
Eg, when created, instances will differ from one proc to the next.

+++ ga/pescan/elemental/evaluator.py
@@ -67,8 +67,10 @@ class Bandgap(object):
  
     # moves to new calculation directory
     self.bandgap.directory = self.directory_prefix + "_" + str(self.nbcalc)
-    if exists(self.bandgap.directory): rmtree( self.bandgap.directory)
-    makedirs(self.bandgap.directory)
+    if self.world.do_print: 
+      if exists(self.bandgap.directory): rmtree( self.bandgap.directory)
+      makedirs(self.bandgap.directory)
+    self.world.barrier()
  
     # moves to calculation directory
     with Changedir(self.bandgap.directory) as pwd:
